moved components and decoupled config
